Refactor searchSong method in SearchSongInteractor
