+++ src/main/java/use_case/playlist_user_story/PlaylistDataAccessInterface.java
@@ -1,4 +1,4 @@
-package use_case.playlist_user_story.playlist;
+package use_case.playlist_user_story;
 
 import entity.Playlist;
 import entity.Song;
