--- src/main/java/use_case/playlist_user_story/search_song/SearchSongDataAccessInterface.java
@@ -1,9 +1,0 @@
-package use_case.playlist_user_story.search_song;
-
-/**
- * Data Access Interface for Search Track.
- */
-public interface SearchSongDataAccessInterface {
-
-
-}
