+++ src/main/java/use_case/playlist_user_story/search_song/SearchSongInteractor.java
@@ -1,10 +1,8 @@
 package use_case.playlist_user_story.search_song;
 
-import entity.Song;
-import se.michaelthelin.spotify.model_objects.specification.ArtistSimplified;
-import se.michaelthelin.spotify.model_objects.specification.Track;
 import data_access.DataAccessException;
-import use_case.playlist_user_story.playlist.PlaylistDataAccessInterface;
+import se.michaelthelin.spotify.model_objects.specification.Track;
+import use_case.playlist_user_story.PlaylistDataAccessInterface;
 
 /**
  * The Interactor for Search Song.
@@ -25,21 +23,8 @@ public class SearchSongInteractor implements SearchSongInputBoundary {
     @Override
     public void searchSong(String query) {
         if (query != null) {
-            final SearchSongOutputData searchSongOutputData = new SearchSongOutputData();
-            final Track[] searchResults = spotifyDAO.searchTrack(query);
-
-            for (Track searchResult : searchResults) {
-                final ArtistSimplified[] artists = searchResult.getArtists();
-                final String[] artistNames = new String[artists.length];
-
-                for (int i = 0; i < artists.length; i++) {
-                    artistNames[i] = artists[i].getName();
-                }
-
-                final Song song = new Song(searchResult.getName(), artistNames);
-                searchSongOutputData.addSong(song);
-            }
-
+            final Track[] tracks = spotifyDAO.searchTrack(query);
+            final SearchSongOutputData searchSongOutputData = new SearchSongOutputData(tracks);
             searchSongPresenter.searchSong(searchSongOutputData);
         }
 
@@ -52,16 +37,13 @@ public class SearchSongInteractor implements SearchSongInputBoundary {
 
     @Override
     public void addSong(SearchSongInputData searchSongInputData) {
-        // TODO: Update playlist in DB with new song
         try {
             playlistDAO.addSongToPlaylist(searchSongInputData.getCurrentPlaylist(),
                     searchSongInputData.getSelectedSong());
 
-            // Update Playlist View with new song
             searchSongPresenter.addSong(searchSongInputData.getSelectedSong());
         }
         catch (DataAccessException exception) {
-            // Prepare some sort of failure message
             searchSongPresenter.prepareFailView(exception.getMessage());
         }
     }
