+++ src/main/java/use_case/playlist_user_story/search_song/SearchSongOutputData.java
@@ -1,5 +1,6 @@
 package use_case.playlist_user_story.search_song;
-
+import se.michaelthelin.spotify.model_objects.specification.ArtistSimplified;
+import se.michaelthelin.spotify.model_objects.specification.Track;
 import entity.Song;
 
 import java.util.ArrayList;
@@ -11,12 +12,18 @@ import java.util.List;
 public class SearchSongOutputData {
     private List<Song> searchResults = new ArrayList<Song>();
 
-    /**
-     * Add Song to searchResults.
-     * @param song the track to be added
-     */
-    public void addSong(Song song) {
-        searchResults.add(song);
+    public SearchSongOutputData(Track[] tracks) {
+        for (Track track : tracks) {
+            final ArtistSimplified[] artists = track.getArtists();
+            final String[] artistNames = new String[artists.length];
+
+            for (int i = 0; i < artists.length; i++) {
+                artistNames[i] = artists[i].getName();
+            }
+
+            final Song song = new Song(track.getName(), artistNames);
+            searchResults.add(song);
+        }
     }
 
     public List<Song> getSearchResults() {
